add test for user story 1
