+++ spec/features/machines/index_spec.rb
