--- spec/features/user_can_see_an_individual_vending_machine_spec.rb
@@ -1,12 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe 'When a user visits a vending machine show page', type: :feature do
-  scenario 'they see the location of that machine' do
-    owner = Owner.create(name: "Sam's Snacks")
-    dons  = owner.machines.create(location: "Don's Mixed Drinks")
-
-    visit machine_path(dons)
-
-    expect(page).to have_content("Don's Mixed Drinks Vending Machine")
-  end
-end
